* origin/master:
  Bug in the fuel actuator Bug in the type of 1D table controller. Bug in the VVT actuators. Bugs in the PID controller
  Fuel mass in petrol engine calculated as a function of fresh air in the cylinder.
  BUG0008 Supersonic flow in boundary conditions.
  BUG0007 Problems with numerical accuracy in 32 bit mingw compiled version
  Initial conditions in the cylinders Species tracking bug in branches solved.
  Correcte the balance of species mass in the cylinder of a 2strokes engine.
  Bugs in gamma in cylinder
  Compressor map extrapolation capabilities.
  Modified corrrelation for gamma in the cylinder.
  Bug in the calculation of the gamma in the cylinder. Bug in the calculation of the instantaneous heat transfer in the cylinder.
  Corrected instantáneous heat to the cylinder walls
  BUG0006: Important memory leak in TCCDesposito.
  BUG0005: Solved different bugs in the TVD, characteristics, etc.
  BUG0004: Important bug in spark ignited engines. The fuel was always 0.
  Add instantaneous heat transfer through the engine walls as a result
  Add OpenWAM icon
  Visual Studio compilation
  BUG0003 Convert pressure and temperature reference for SAE compressor map: C -> K and bar -> Pa

+++ Source/Connections/TLamina.h
@@ -63,7 +63,7 @@ double FAnchoPetalo;
 double FRigidez;
 double FModuloYoung;
 double FEspesor;
-double FNumPestanyas;
+int FNumPestanyas;
 double FLongitud;
 double FLongReal;
